Finished Sketch 17 - What if Pi equals 4?
I didn't really dive too deep into this prompt, I just replaced Pi in my spiral position generator to get a different pattern than the Fibonacci sequence. I created a new macro library to hold some new utility operations, starting with "for loop with delay" and "for each loop with delay" so that I could get the behavior I was looking for with spawning pooled actors. Then I thought it would be fun to make tentacles, so I made a new pooled actor class and created a new Niagara system of a ribbon renderer with a procedurally generated tentacle material